src/main/kotlin/no/nav/eessi/pensjon/services/kodeverk/KodeverkClient.kt
package no.nav.eessi.pensjon.services.kodeverk

import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
import io.micrometer.core.instrument.simple.SimpleMeterRegistry
import no.nav.eessi.pensjon.metrics.MetricsHelper
import no.nav.eessi.pensjon.utils.toJson
import org.slf4j.LoggerFactory
import org.springframework.beans.factory.annotation.Autowired
import org.springframework.beans.factory.annotation.Value
import org.springframework.cache.annotation.CacheConfig
import org.springframework.cache.annotation.Cacheable
import org.springframework.http.HttpEntity
import org.springframework.http.HttpHeaders
import org.springframework.http.HttpMethod
import org.springframework.http.HttpStatus
import org.springframework.stereotype.Component
import org.springframework.web.bind.annotation.ResponseStatus
import org.springframework.web.client.HttpClientErrorException
import org.springframework.web.client.HttpServerErrorException
import org.springframework.web.client.RestTemplate
import org.springframework.web.util.UriComponents
import org.springframework.web.util.UriComponentsBuilder
import java.util.*
import javax.annotation.PostConstruct

@Component
@CacheConfig(cacheNames = ["kodeVerk"])
class KodeverkClient(private val kodeRestTemplate: RestTemplate,
                     @Value("\${NAIS_APP_NAME}") private val appName: String,
                     @Autowired(required = false) private val metricsHelper: MetricsHelper = MetricsHelper(SimpleMeterRegistry())) {

    private val logger = LoggerFactory.getLogger(KodeverkClient::class.java)

    private lateinit var KodeverkHentLandKode: MetricsHelper.Metric

    @PostConstruct
    fun initMetrics() {
        KodeverkHentLandKode = metricsHelper.init("KodeverkHentLandKode")
    }

    fun hentAlleLandkoder() = hentLandKoder().toJson()

    fun hentLandkoderAlpha2() = hentLandKoder().map { it.landkode2 }.toList()

    @Cacheable
    fun hentLandKoder(): List<Landkode> {
        return KodeverkHentLandKode.measure {
            val tmpLandkoder = hentHierarki("LandkoderSammensattISO2")

            val rootNode = jacksonObjectMapper().readTree(tmpLandkoder)
            val noder = rootNode.at("/noder").toList()

            noder.map { node ->
                Landkode(node.at("/kode").textValue(),
                        node.at("/undernoder").findPath("kode").textValue())
            }.sortedBy { (sorting, _) -> sorting }.toList()
        }
    }

    @Cacheable("string")
    fun finnLandkode2(alpha3: String): String? {
        val list = hentLandKoder()
        list.forEach {
            if (it.landkode3 == alpha3) {
                return it.landkode2
            }
        }
        return null
    }

    @Cacheable("string")
    fun finnLandkode3(alpha2: String): String? {
        val list = hentLandKoder()
        list.forEach {
            if (it.landkode2 == alpha2) {
                return it.landkode3
            }
        }
        return null

    }

    private fun doRequest(builder: UriComponents) : String {
        return try {
            val headers = HttpHeaders()
            headers["Nav-Consumer-Id"] = appName
            headers["Nav-Call-Id"] = UUID.randomUUID().toString()
            val requestEntity = HttpEntity<String>(headers)
            logger.debug("Header: $requestEntity")
            val response = kodeRestTemplate.exchange(
                    builder.toUriString(),
                    HttpMethod.GET,
                    requestEntity,
                    String::class.java)

            response.body?.let { return it } ?: {
                throw KodeverkException("Feil ved konvetering av jsondata fra kodeverk")
            }()

        } catch (ce: HttpClientErrorException) {
            logger.error(ce.message, ce)
            throw KodeverkException(ce.message!!)
        } catch (se: HttpServerErrorException) {
            logger.error(se.message, se)
            throw KodeverkException(se.message!!)
        } catch (ex: Exception) {
            logger.error(ex.message, ex)
            throw KodeverkException(ex.message!!)
        }
    }

    /**
     *  https://kodeverk.nais.adeo.no/api/v1/hierarki/LandkoderSammensattISO2/noder
     */
    private fun hentHierarki(hierarki: String) : String {
        val path = "/api/v1/hierarki/{hierarki}/noder"

        val uriParams = mapOf("hierarki" to hierarki)
        val builder = UriComponentsBuilder.fromUriString(path).buildAndExpand(uriParams)

        return doRequest(builder)
    }
}

data class Landkode (
        val landkode2: String, // SE
        val landkode3: String // SWE
)

@ResponseStatus(value = HttpStatus.INTERNAL_SERVER_ERROR)
class KodeverkException(message: String) : RuntimeException(message)
